Moved fixtures to conftest
