+++ pvfactors/tests/conftest.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from pvfactors.pvrow import PVRowSide
+from pvfactors.pvsurface import PVSegment, PVSurface
 import pytest
 import os
 import pandas as pd
@@ -110,3 +112,21 @@ def df_registries_with_nan(df_registries):
                                        axis=0, sort=False)
 
     yield df_registries_with_nan
+
+
+######################################################################
+
+
+@pytest.fixture(scope='function')
+def pvsegments():
+    seg_1 = PVSegment(
+        illum_surface=PVSurface([(0, 0), (1, 0)], shaded=False))
+    seg_2 = PVSegment(
+        shaded_surface=PVSurface([(1, 0), (2, 0)], shaded=True))
+    yield seg_1, seg_2
+
+
+@pytest.fixture(scope='function')
+def pvrow_side(pvsegments):
+    side = PVRowSide(pvsegments)
+    yield side
